Add filter field and home component
